move epoch info
