+++ chain/chunks/src/client.rs
@@ -7,7 +7,7 @@ use near_pool::types::TransactionGroupIterator;
 use near_pool::{InsertTransactionResult, PoolIteratorWrapper, TransactionPool};
 use near_primitives::shard_layout::{account_id_to_shard_uid, ShardLayout, ShardUId};
 use near_primitives::{
-    epoch_manager::RngSeed,
+    epoch_info::RngSeed,
     sharding::{EncodedShardChunk, PartialEncodedChunk, ShardChunk, ShardChunkHeader},
     transaction::SignedTransaction,
     types::{AccountId, ShardId},
@@ -157,7 +157,7 @@ mod tests {
     use near_o11y::testonly::init_test_logger;
     use near_pool::types::TransactionGroupIterator;
     use near_primitives::{
-        epoch_manager::RngSeed,
+        epoch_info::RngSeed,
         hash::CryptoHash,
         shard_layout::{account_id_to_shard_uid, ShardLayout},
         transaction::SignedTransaction,
