+++ chain/client/src/test_utils/test_env_builder.rs
@@ -16,7 +16,8 @@ use near_epoch_manager::shard_tracker::{ShardTracker, TrackedConfig};
 use near_epoch_manager::{EpochManager, EpochManagerAdapter, EpochManagerHandle};
 use near_network::test_utils::MockPeerManagerAdapter;
 use near_parameters::RuntimeConfigStore;
-use near_primitives::epoch_manager::{AllEpochConfigTestOverrides, RngSeed};
+use near_primitives::epoch_info::RngSeed;
+use near_primitives::epoch_manager::AllEpochConfigTestOverrides;
 use near_primitives::test_utils::create_test_signer;
 use near_primitives::types::{AccountId, NumShards};
 use near_store::config::StateSnapshotType;
