+++ chain/epoch-manager/src/tests/random_epochs.rs
@@ -8,8 +8,8 @@ use crate::test_utils::{
 };
 use crate::EpochManager;
 use near_primitives::challenge::SlashedValidator;
+use near_primitives::epoch_info::EpochInfo;
 use near_primitives::epoch_manager::block_info::BlockInfo;
-use near_primitives::epoch_manager::epoch_info::EpochInfo;
 use near_primitives::epoch_manager::SlashState;
 use near_primitives::hash::CryptoHash;
 use near_primitives::types::validator_stake::ValidatorStake;
