+++ chain/epoch-manager/src/types.rs
@@ -1,8 +1,8 @@
 use borsh::{BorshDeserialize, BorshSerialize};
 use near_primitives::block_header::BlockHeader;
 use near_primitives::challenge::SlashedValidator;
+use near_primitives::epoch_info::EpochInfo;
 use near_primitives::epoch_manager::block_info::BlockInfo;
-use near_primitives::epoch_manager::epoch_info::EpochInfo;
 use near_primitives::hash::CryptoHash;
 use near_primitives::types::validator_stake::ValidatorStake;
 use near_primitives::types::{
