+++ core/primitives/src/epoch_sync.rs
@@ -1,6 +1,6 @@
 use crate::block_header::BlockHeader;
+use crate::epoch_info::EpochInfo;
 use crate::epoch_manager::block_info::BlockInfo;
-use crate::epoch_manager::epoch_info::EpochInfo;
 use crate::merkle::PartialMerkleTree;
 use crate::views::LightClientBlockView;
 use borsh::{BorshDeserialize, BorshSerialize};
