+++ core/store/src/columns.rs
@@ -71,7 +71,7 @@ pub enum DBCol {
     RecentOutboundConnections,
     /// Mapping from EpochId to EpochInfo
     /// - *Rows*: EpochId (CryptoHash)
-    /// - *Content type*: [near_primitives::epoch_manager::epoch_info::EpochInfo]
+    /// - *Content type*: [near_primitives::epoch_info::EpochInfo]
     EpochInfo,
     /// Mapping from BlockHash to BlockInfo
     /// - *Rows*: BlockHash (CryptoHash)
