+++ core/store/src/migrations.rs
@@ -1,7 +1,7 @@
 use crate::metadata::DbKind;
 use crate::{DBCol, Store, StoreUpdate};
 use borsh::{BorshDeserialize, BorshSerialize};
-use near_primitives::epoch_manager::epoch_info::EpochSummary;
+use near_primitives::epoch_manager::EpochSummary;
 use near_primitives::epoch_manager::AGGREGATOR_KEY;
 use near_primitives::hash::CryptoHash;
 use near_primitives::state::FlatStateValue;
