+++ tools/protocol-schema-check/res/protocol_schema.toml
@@ -50,8 +50,12 @@ EncodedShardChunkBody = 2481614037
 EncodedShardChunkV1 = 3000186041
 EncodedShardChunkV2 = 3338195984
 EpochId = 1173955846
+EpochInfo = 3439987361
 EpochInfoV1 = 760560610
+EpochInfoV2 = 2533281205
+EpochInfoV3 = 91327628
 EpochInfoV4 = 1127098481
+EpochSummary = 3803241212
 EpochValidatorInfo = 378323971
 ExecutionMetadata = 3532864429
 ExecutionOutcome = 1797422966
