+++ tools/state-viewer/src/scan_db.rs
@@ -2,8 +2,8 @@ use borsh::BorshDeserialize;
 use near_chain::types::LatestKnown;
 use near_epoch_manager::types::EpochInfoAggregator;
 use near_primitives::block::{Block, BlockHeader, Tip};
+use near_primitives::epoch_info::EpochInfo;
 use near_primitives::epoch_manager::block_info::BlockInfo;
-use near_primitives::epoch_manager::epoch_info::EpochInfo;
 use near_primitives::epoch_manager::AGGREGATOR_KEY;
 use near_primitives::receipt::Receipt;
 use near_primitives::shard_layout::{get_block_shard_uid_rev, ShardUId};
